handle download image with rxswift
