+++ RxSwift_Template/View/Controllers/Home/Cell/Popular/PopularCollection/PopularCollectionViewCell.swift
@@ -41,19 +41,8 @@ final class PopularCollectionViewCell: UICollectionViewCell {
             .disposed(by: bag)
         
         popular.map(\.imageProduct)
-            .bind(to: productImageView.rx.downloadImage)
+            .flatMap { DownloadImage.shared.dowloadImageWithRxSwift(url: $0 ?? "") }
+            .bind(to: productImageView.rx.image)
             .disposed(by: bag)
-    }
-}
-
-extension Reactive where Base: UIImageView {
-
-    var downloadImage: Binder<String?> {
-        return Binder(self.base) { imageView, stringImage in
-            UIImageView.dowloadImageWithRxSwift(url: stringImage ?? "")
-                .subscribe { image in
-                    imageView.rx.image.onNext(image)
-                }
-        }
     }
 }
