+++ RxSwift_Template/View/Controllers/Home/Cell/Recommend/RecommendCollection/RecommendCollectionViewCell.swift
@@ -40,9 +40,10 @@ final class RecommendCollectionViewCell: UICollectionViewCell {
         recommemd.map(\.category?.nameCategory)
             .bind(to: priceProductLabel.rx.text)
             .disposed(by: bag)
-        
+
         recommemd.map(\.imageProduct)
-            .bind(to: productImageView.rx.downloadImage)
+            .flatMap { DownloadImage.shared.dowloadImageWithRxSwift(url: $0 ?? "") }
+            .bind(to: productImageView.rx.image)
             .disposed(by: bag)
     }
 }
