+++ RxSwift_Template/View/Controllers/Home/Cell/Slider/SlideCollection/SlideCollectionViewCell.swift
@@ -32,7 +32,8 @@ final class SlideCollectionViewCell: UICollectionViewCell {
             .disposed(by: bag)
         
         shop.map(\.imageShop)
-            .bind(to: shopImageView.rx.downloadImage)
+            .flatMap { DownloadImage.shared.dowloadImageWithRxSwift(url: $0 ?? "") }
+            .bind(to: shopImageView.rx.image)
             .disposed(by: bag)
     }
 }
